% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_toydata.R
\name{create_toydata}
\alias{create_toydata}
\title{create_toydata}
\usage{
create_toydata(n = 1000, seed = NULL)
}
\arguments{
\item{n}{The number of observations in the produced toydata}

\item{seed}{Seed to get reproducible data}
}
\value{
A matrix with two columns and n rows.
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
\examples{
toydata <- create_toydata(n = 1000)
}
