% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_toydata.R
\name{plot_toydata}
\alias{plot_toydata}
\title{plot_toydata}
\usage{
plot_toydata(dat, col = viridis::viridis(2, alpha = 0.7)[1], pch = 19, ...)
}
\arguments{
\item{dat}{The number of observations in the produced toydata}

\item{col}{Set your favorite color (if you don't like the default)}

\item{...}{Additional arguments passed to plot}
}
\value{
A plot of the toydata
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
\examples{
toydata <- create_toydata(n = 1000)
plot_toydata(toydata)
}
